# Separate InternalTypes to allow integer operation optimizations