--- a/Jint/Key.cs
+++ b/Jint/Key.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Diagnostics;
+using System.Runtime.CompilerServices;
 using Jint.Runtime;
 
 namespace Jint
@@ -11,6 +12,17 @@ namespace Jint
     [DebuggerDisplay("{" + nameof(Name) + "}")]
     public readonly struct Key : IEquatable<Key>
     {
+        // lookup for indexer keys
+        internal static readonly Key[] indexKeys = new Key[TypeConverter.intToString.Length];
+
+        static Key()
+        {
+            for (uint i = 0; i < indexKeys.Length; ++i)
+            {
+                indexKeys[i] = new Key(TypeConverter.ToString(i));
+            }
+        }
+
         public Key(string name)
         {
             if (name == null)
@@ -27,6 +39,24 @@ namespace Jint
         public static implicit operator Key(string name)
         {
             return new Key(name);
+        }
+
+        public static implicit operator Key(int value)
+        {
+            var keys = indexKeys;
+            return (uint) value < keys.Length ? keys[value] : BuildKey(value);
+        }
+
+        public static implicit operator Key(uint value)
+        {
+            var keys = indexKeys;
+            return value < keys.Length ? keys[value] : BuildKey(value);
+        }
+
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private static Key BuildKey(long value)
+        {
+            return new Key(value.ToString());
         }
 
         public static implicit operator string(Key key) => key.Name;