--- a/Jint/Native/Array/ArrayPrototype.cs
+++ b/Jint/Native/Array/ArrayPrototype.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Runtime.CompilerServices;
 using Jint.Collections;
 using Jint.Native.Number;
 using Jint.Native.Object;
@@ -1321,15 +1322,22 @@ namespace Jint.Native.Array
 
                 public override ObjectInstance Target => _instance;
 
-                internal double GetIntegerLength()
+                [MethodImpl(MethodImplOptions.AggressiveInlining)]
+                internal ulong GetIntegerLength()
                 {
                     var desc = _instance.GetProperty("length");
                     var descValue = desc.Value;
                     if (desc.IsDataDescriptor() && !ReferenceEquals(descValue, null))
                     {
-                        return TypeConverter.ToInteger(descValue);
+                        return TypeConverter.ToUnsignedInteger(descValue);
                     }
 
+                    return GetIntegerLengthUnlikely(desc);
+                }
+
+                [MethodImpl(MethodImplOptions.NoInlining)]
+                private ulong GetIntegerLengthUnlikely(PropertyDescriptor desc)
+                {
                     var getter = desc.Get ?? Undefined;
                     if (getter.IsUndefined())
                     {
@@ -1339,7 +1347,7 @@ namespace Jint.Native.Array
                     // if getter is not undefined it must be ICallable
                     var callable = (ICallable) getter;
                     var value = callable.Call(_instance, Arguments.Empty);
-                    return TypeConverter.ToInteger(value);
+                    return TypeConverter.ToUnsignedInteger(value);
                 }
 
                 public override ulong GetSmallestIndex(ulong length)
@@ -1375,18 +1383,12 @@ namespace Jint.Native.Array
 
                 public override uint GetLength()
                 {
-                    var integerLength = GetIntegerLength();
-                    return (uint) (integerLength >= 0 ? integerLength : 0);
+                    return (uint) GetIntegerLength();
                 }
 
                 public override ulong GetLongLength()
                 {
-                    var integerLength = GetIntegerLength();
-                    if (integerLength <= 0)
-                    {
-                        return 0;
-                    }
-                    return (ulong) System.Math.Min(integerLength, MaxArrayLikeLength);
+                    return System.Math.Min(GetIntegerLength(), MaxArrayLikeLength);
                 }
 
                 public override void SetLength(ulong length) => _instance.Put("length", length, true);