--- a/Jint/Native/Function/ScriptFunctionInstance.cs
+++ b/Jint/Native/Function/ScriptFunctionInstance.cs
@@ -11,6 +11,8 @@ namespace Jint.Native.Function
     public sealed class ScriptFunctionInstance : FunctionInstance, IConstructor
     {
         internal readonly JintFunctionDefinition _function;
+        private LexicalEnvironment _localEnv;
+
 
         /// <summary>
         /// http://www.ecma-international.org/ecma-262/5.1/#sec-13.2
@@ -86,7 +88,9 @@ namespace Jint.Native.Function
                     thisBinding = thisArg;
                 }
 
-                var localEnv = LexicalEnvironment.NewDeclarativeEnvironment(_engine, _scope);
+                var localEnv = _localEnv ?? LexicalEnvironment.NewDeclarativeEnvironment(_engine, _scope);
+                localEnv.Reset(_scope);
+                _localEnv = null;
 
                 _engine.EnterExecutionContext(localEnv, localEnv, thisBinding);
 
@@ -121,6 +125,7 @@ namespace Jint.Native.Function
                 finally
                 {
                     _engine.LeaveExecutionContext();
+                    _localEnv = localEnv;
                 }
 
                 return Undefined;