--- a/Jint/Native/String/StringPrototype.cs
+++ b/Jint/Native/String/StringPrototype.cs
@@ -848,7 +848,7 @@ namespace Jint.Native.String
             {
                 return JsNumber.DoubleNaN;
             }
-            return (double) s[position];
+            return (long) s[position];
         }
 
         private JsValue CodePointAt(JsValue thisObj, JsValue[] arguments)
@@ -863,10 +863,10 @@ namespace Jint.Native.String
                 return Undefined;
             }
 
-            var first = (double) s[position];
+            var first = (long) s[position];
             if (first >= 0xD800 && first <= 0xDBFF && s.Length > position + 1)
             {
-                double second = s[position + 1];
+                long second = s[position + 1];
                 if (second >= 0xDC00 && second <= 0xDFFF)
                 {
                     return (first - 0xD800) * 0x400 + second - 0xDC00 + 0x10000;
@@ -886,7 +886,6 @@ namespace Jint.Native.String
                 return "";
             }
             return TypeConverter.ToString(s[(int) position]);
-
         }
 
         private JsValue ValueOf(JsValue thisObj, JsValue[] arguments)