… Jint/Runtime/Environments/LexicalEnvironment.cs
… Jint/Runtime/Interpreter/Expressions/JintBinaryExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintCallExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintConstantExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintLiteralExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintMemberExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintUnaryExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintUpdateExpression.cs
… Jint/Runtime/Interpreter/Statements/JintForInStatement.cs
… Jint/Runtime/References/Reference.cs
… Jint/Runtime/TypeConverter.cs