Here I'm trying out using separate known flag to safe integers. When we know it's a safe integer, we can skip both floating point arithmetic and use faster lookups for results. Consumers still see the exposed `Types` but internally we handle with `InternalTypes` which also has `Integer`.