@lahma 2019-10-16
Results for the first version, probably need to go through with a finer comb still...

## Esprima.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|
|------- |-------|-------|-------:|
| Old |Run|3d-cube|441.7 ms|
| **New** |	|	| **404.2 ms (-8%)** |
| Old |Run|3d-morph|401.7 ms|
| **New** |	|	| **360.4 ms (-10%)** |
| Old |Run|3d-raytrace|NA|
| **New** |	|	| **NA** |
| Old |Run|access-binary-trees|160.5 ms|
| **New** |	|	| **155.9 ms (-3%)** |
| Old |Run|access-fannkuch|1|
| **New** |	|	| **1 (0%)** |
| Old |Run|access-nbody|366.0 ms|
| **New** |	|	| **354.8 ms (-3%)** |
| Old |Run|access-nsieve|462.5 ms|
| **New** |	|	| **412.4 ms (-11%)** |
| Old |Run|bitop(...)-byte [24]|277.7 ms|
| **New** |	|	| **259.1 ms (-7%)** |
| Old |Run|bitops-bits-in-byte|448.9 ms|
| **New** |	|	| **400.1 ms (-11%)** |
| Old |Run|bitops-bitwise-and|225.0 ms|
| **New** |	|	| **209.9 ms (-7%)** |
| Old |Run|bitops-nsieve-bits|448.1 ms|
| **New** |	|	| **NA** |
| Old |Run|contr(...)rsive [21]|229.0 ms|
| **New** |	|	| **219.5 ms (-4%)** |
| Old |Run|crypto-aes|320.7 ms|
| **New** |	|	| **268.8 ms (-16%)** |
| Old |Run|crypto-md5|221.3 ms|
| **New** |	|	| **207.4 ms (-6%)** |
| Old |Run|crypto-sha1|211.7 ms|
| **New** |	|	| **205.5 ms (-3%)** |
| Old |Run|date-format-tofte|258.9 ms|
| **New** |	|	| **242.3 ms (-6%)** |
| Old |Run|date-format-xparb|256.3 ms|
| **New** |	|	| **254.5 ms (-1%)** |
| Old |Run|math-cordic|656.7 ms|
| **New** |	|	| **613.5 ms (-7%)** |
| Old |Run|math-partial-sums|223.2 ms|
| **New** |	|	| **213.1 ms (-5%)** |
| Old |Run|math-spectral-norm|247.0 ms|
| **New** |	|	| **225.7 ms (-9%)** |
| Old |Run|regexp-dna|342.2 ms|
| **New** |	|	| **344.7 ms (+1%)** |
| Old |Run|string-base64|189.3 ms|
| **New** |	|	| **184.8 ms (-2%)** |
| Old |Run|string-fasta|340.7 ms|
| **New** |	|	| **328.6 ms (-4%)** |
| Old |Run|string-tagcloud|155.5 ms|
| **New** |	|	| **154.8 ms (0%)** |
| Old |Run|string-unpack-code|146.8 ms|
| **New** |	|	| **147.6 ms (+1%)** |
| Old |Run|strin(...)input [21]|138.2 ms|
| **New** |	|	| **136.5 ms (-1%)** |


## Jint.Benchmark.ArrayBenchmark

| **Diff**|Method|N|Mean|
|------- |-------|-------|-------:|
| Old |Slice|100|373.3 us|
| **New** |	|	| **365.4 us (-2%)** |
| Old |Concat|100|508.3 us|
| **New** |	|	| **502.6 us (-1%)** |
| Old |Unshift|100|18|
| **New** |	|	| **16 (-11%)** |
| Old |Push|100|7|
| **New** |	|	| **6 (-14%)** |
| Old |Index|100|9|
| **New** |	|	| **7 (-22%)** |
| Old |Map|100|3|
| **New** |	|	| **3 (0%)** |
| Old |Apply|100|589.1 us|
| **New** |	|	| **591.0 us (0%)** |
| Old |JsonStringifyParse|100|4|
| **New** |	|	| **4 (0%)** |
| Old |FilterWithString|100|35|
| **New** |	|	| **33 (-6%)** |


## Jint.Benchmark.ArrayStressBenchmark

| **Diff**|Method|N|Mean|
|------- |-------|-------|-------:|
| Old |Jint|20|446.1 ms|
| **New** |	|	| **392.4 ms (-12%)** |


## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|
|------- |-------|-------|-------:|
| Old |Run|dromaeo-3d-cube|56.42 ms|
| **New** |	|	| **51.43 ms (-9%)** |
| Old |Run|dromaeo-core-eval|13.77 ms|
| **New** |	|	| **11.98 ms (-13%)** |
| Old |Run|dromaeo-object-array|133.80 ms|
| **New** |	|	| **125.55 ms (-6%)** |
| Old |Run|droma(...)egexp [21]|591.96 ms|
| **New** |	|	| **569.59 ms (-4%)** |
| Old |Run|droma(...)tring [21]|478.53 ms|
| **New** |	|	| **474.61 ms (-1%)** |


## Jint.Benchmark.EngineConstructionBenchmark

| **Diff**|Method|Mean|
|------- |-------|-------:|
| Old |BuildEngine|2.506 us|
| **New** |	| **2.443 us (-3%)** |


## Jint.Benchmark.EvaluationBenchmark

| **Diff**|Method|N|Mean|
|------- |-------|-------|-------:|
| Old |Jint|20|883.4 us|
| **New** |	|	| **876.4 us (-1%)** |


## Jint.Benchmark.InteropBenchmark

| **Diff**|Method|Mean|
|------- |-------|-------:|
| Old |DelegatesCanBeSet|7.655 us|
| **New** |	| **8.170 us (+7%)** |
| Old |ExtraParametersAreIgnored|12.374 us|
| **New** |	| **12.171 us (-2%)** |
| Old |GetObjectProperties|6.177 us|
| **New** |	| **6.183 us (0%)** |
| Old |InvokeObjectMethods|7.224 us|
| **New** |	| **7.261 us (+1%)** |
| Old |SetObjectProperties|9.334 us|
| **New** |	| **9.359 us (0%)** |
| Old |GetIndexUsingStringKey|19.005 us|
| **New** |	| **19.646 us (+3%)** |
| Old |GenericMethods|21.042 us|
| **New** |	| **21.097 us (0%)** |
| Old |MultiGenericTypes|46.543 us|
| **New** |	| **46.390 us (0%)** |
| Old |IndexOnList|12.855 us|
| **New** |	| **12.942 us (+1%)** |
| Old |EcmaValuesAreAutomaticallyConvertedWhenSetInPoco|17.238 us|
| **New** |	| **17.001 us (-1%)** |
| Old |Trim|16.128 us|
| **New** |	| **15.934 us (-1%)** |
| Old |MathFloor|10.499 us|
| **New** |	| **11.052 us (+5%)** |
| Old |DelegateAsFunction|7.897 us|
| **New** |	| **7.778 us (-2%)** |
| Old |ConvertArrayToArrayInstance|13.177 us|
| **New** |	| **13.061 us (-1%)** |
| Old |ConvertListsToArrayInstance|15.568 us|
| **New** |	| **16.025 us (+3%)** |
| Old |ConvertArrayInstanceToArray|3.701 us|
| **New** |	| **3.803 us (+3%)** |
| Old |LoopWithNativeEnumerator|7.885 us|
| **New** |	| **8.129 us (+3%)** |


## Jint.Benchmark.LinqJsBenchmark

| **Diff**|Method|N|Mean|
|------- |-------|-------|-------:|
| Old |Jint|10|66.25 ms|
| **New** |	|	| **66.72 ms (+1%)** |


## Jint.Benchmark.MinimalScriptBenchmark

| **Diff**|Method|N|Mean|
|------- |-------|-------|-------:|
| Old |Jint|10|25.84 us|
| **New** |	|	| **26.28 us (+2%)** |


## Jint.Benchmark.StopwatchBenchmark

| **Diff**|Method|N|Mean|
|------- |-------|-------|-------:|
| Old |Jint|1|810.0 ms|
| **New** |	|	| **920.7 ms (+14%)** |


## Jint.Benchmark.StringBuilderBenchmark

| **Diff**|Method|Mean|
|------- |-------|-------:|
| Old |One|2.266 us|
| **New** |	| **2.292 us (+1%)** |
| Old |Two|2.996 us|
| **New** |	| **3.039 us (+1%)** |
| Old |Three|3.708 us|
| **New** |	| **3.668 us (-1%)** |


## Jint.Benchmark.UncacheableExpressionsBenchmark

| **Diff**|Method|N|Mean|
|------- |-------|-------|-------:|
| Old |Benchmark|500|231.2 ms|
| **New** |	|	| **237.3 ms (+3%)** |