@lahma 2019-10-16
Yep, floating point can make things really slow.  I think I'll have the things I wanted to do ready quite soon. Can utilize internal types later on more.