@lahma 2019-10-18
There's still a problem with `bitops-nsieve-bits.js`  that I need to resolve, so not ready yet..